Bugfix with types in cdefs.h

+++ bmp/cdefs.h
@@ -1,13 +1,14 @@
 #ifndef _CDEFS_
 #define _CDEFS_
+#include <stdint.h>
 
 typedef unsigned char  uchar;      // беззнаковый однобайтовый символ
 typedef   signed char  schar;      // знаковый однобайтовый символ
 typedef unsigned int   uint;       // беззнаковое целое
 
-typedef   signed short int16;      // знаковое 16-битное целое
-typedef unsigned short uint16;     // беззнаковое 16-битное целое
-typedef   signed long  int32;      // знаковое 32-битное целое
-typedef unsigned long  uint32;     // беззнаковое 32-битное целое
+typedef int16_t  int16;      // знаковое 16-битное целое
+typedef uint16_t uint16;     // беззнаковое 16-битное целое
+typedef int32_t  int32;      // знаковое 32-битное целое
+typedef uint32_t uint32;     // беззнаковое 32-битное целое
 
 #endif // _CDEFS_
